Implement rudimentary first step of the RG algorithm.

+++ src-tauri/src/common.rs
@@ -11,15 +11,15 @@ pub(crate) fn find_active_transitions(marking: &Array1<i32>, transition_inputs: 
     return active;
 }
 
-pub(crate) fn find_active_transitions_arr(marking: &Array1<u32>, transition_inputs: &Array2<i32>) -> Vec<u32> {
+pub(crate) fn find_active_transitions_arr(marking: &Array1<i32>, transition_inputs: &Array2<i32>) -> Vec<i32> {
     let mut active_transitions = Vec::new();
 
     // Compare each row of the matrix to the reference array
     for (row_index, row) in transition_inputs.axis_iter(ndarray::Axis(0)).enumerate() {
         // Check whether the marking is at least as large as the edge weight.
-        if marking.iter().zip(row.iter()).all(|(&a, &b)| a >= b as u32) {
+        if marking.iter().zip(row.iter()).all(|(&a, &b)| a >= b) {
             // If the marking is large enough, the transition is active.
-            active_transitions.push(row_index as u32);
+            active_transitions.push(row_index as i32);
         }
     }
